// packages/splits-sdk/src/subgraphv2/types.ts
type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  DateTime: string
}

export type GqlRecipient = {
  id: Scalars['ID']
  account: GqlAccount
  ownership: Scalars['String']
}

type GqlAccountSharedFields = {
  chainId: Scalars['String']
  distributions: GqlTokenBalance[]
  latestBlock: Scalars['Int']
  latestActivity: Scalars['String']
  parentEntityType?: Scalars['String']
  internalBalances: GqlTokenBalance[]
  warehouseBalances: GqlTokenBalance[]
  contractEarnings: GqlContractEarnings[]
  warehouseWithdrawConfig: GqlWarehouseWithdrawConfig
}
// ...
type GqlContractEarningsBalance = GqlTokenBalanceSharedData & {
  contract: GqlAccount
}

export type GqlContractEarnings = {
  contract: GqlAccount
  internalBalances: GqlContractEarningsBalance[]
  withdrawals: GqlContractEarningsBalance[]
}

export type GqlUser = GqlAccountSharedFields & {
  __typename: 'User'
  id: Scalars['ID']
  withdrawals: GqlTokenBalance[]
  contractEarnings: GqlContractEarnings[]
}
// ...
export type GqlTokenBalanceSharedData = {
  id: Scalars['ID']
  amount: Scalars['Int']
}

export type GqlTokenBalance = GqlTokenBalanceSharedData & {
  id: Scalars['ID']
  account: GqlAccount
}

export type GqlWarehouseWithdrawConfig = {
  paused: Scalars['Boolean']
  incentive: Scalars['Int']
}
// ...
export type GqlAccount =
  | GqlUser
  | GqlSplit
  | GqlVestingModule
  | GqlWaterfallModule
  | GqlLiquidSplit
  | GqlSwapper
  | GqlPassThroughWallet

export type GqlAccountType = GqlAccount['__typename']

export type IAddress = `0x${string}`
export type IBalance = Dictionary<bigint>
export type ISplitType = 'immutable' | 'mutable' | 'liquid'

export type IAccount = {
  address: IAddress
  chainId: SupportedChainId
}

type IAccountSharedFields = IAccount & {
  distributions: IBalance
  latestBlock: number
  latestActivity: number
  parentEntityType?: string
  controllingSplits?: IAddress[]
  pendingControlSplits?: IAddress[]
  ownedSwappers?: IAddress[]
  ownedPassThroughWallets?: IAddress[]
  upstreamSplits?: IAddress[]
  upstreamWaterfalls?: IAddress[]
  upstreamLiquidSplits?: IAddress[]
  upstreamSwappers?: IAddress[]
  upstreamPassThroughWallets?: IAddress[]
  upstreamVesting?: IAddress[]
  contractEarnings: { [address: string]: IBalance }
  warehouseWithdrawConfig?: {
    paused: boolean
    incentive: number
  }
}
// ...
export type IUser = IAccountSharedFields & {
  type: 'user'
  withdrawn: IBalance
  balances: IBalance
  warehouseBalances: IBalance
  contractEarnings: { [address: string]: IBalance }
}